refactor: view 관련 파일들 views 폴더로 이동
